requirements.txt
numpy>=1.17
matplotlib>=3.0.0
pyglet==1.3.2
PyVirtualDisplay==0.2.1
gym[atari,box2d]==0.10.9
nn_builder
scipy
tensorboard
pyyaml
tqdm
pandas
sklearn
imageio
seaborn
pyflann
2to3
pyarrow
cloudpickle
gensim
pybind11
